requirements.txt
numpy==1.26.3
opencv-python==4.9.0.80
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pyzbar==0.1.9
